cdc: Move "options" to separate header to avoid to much header inclusion
cdc should not contaminate the whole universe.
